+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_iscsi_security.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2019-2025, NetApp, Inc
+# (c) 2019-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 '''
@@ -65,6 +65,30 @@ options:
       - Outbound CHAP user password.
       - Can not be modified. If want to change password, delete and re-create the initiator.
     type: str
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 short_description: "NetApp ONTAP Manage iscsi security."
 version_added: "19.11.0"
 '''
@@ -137,12 +161,14 @@ class NetAppONTAPIscsiSecurity:
             outbound_password=dict(required=False, type='str', no_log=True),
             outbound_username=dict(required=False, type='str'),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             required_if=[
-                ['auth_type', 'chap', ['inbound_username', 'inbound_password']]
+                ['auth_type', 'chap', ['inbound_username', 'inbound_password']],
+                ['use_lambda', True, ('lambda_config',)]
             ],
             required_together=[
                 ['inbound_username', 'inbound_password'],